fix: confirmation footer component, delete refresh handling
